WEB-4460 move PatientLastReviewed to a generic adapter in PatientLastReviewed
